Use isolated scratch space
This commit makes it possible to run scripts tracked by BT in parallel
with each other.  That has at least two use cases -- running multiple
scripts (e.g., test suites) in parallel, and sharing a build server with
other people who also use BT.

The prior behavior was to put scratch files in hard-coded names in `/tmp`,
like `/tmp/bt.START` and `/tmp/bt.CPU`.  The new behavior is to create a
scratch directory with `mktemp` and put all the files in there, like
`/tmp/bt-12345-asdfasd/START`.  The scratch directory is exported in the
variable `BT_DIR`, so subprocesses know where to find it.

In addition to allowing parallel execution, this is a modest security
improvement.  Creating and writing files with predictable names in
world-writable directories like `/tmp` allows an attacker with access to
the same host to clobber files owned by you.